+++ WebAnalytics.Core/DTOs/PageSpeedData.cs
@@ -10,7 +10,7 @@ namespace WebAnalytics.Core.DTOs
     {
         public string Date { get; set; } = string.Empty;
         public string Page { get; set; } = string.Empty;
-        public decimal PerformanceScore { get; set; }
+        public double PerformanceScore { get; set; }
         public int LCP_ms { get; set; }
     }
 }
